refactor: Move sql functions to one file
